+++ shabunkov.kirill/T3/commands.h
@@ -19,7 +19,7 @@ double areaNumOfVertex(size_t vertex, const std::vector<Polygon>& polys);
 double maxAreaVertexes(const std::string& parametr, const std::vector<Polygon>& polys);
 double minAreaVertexes(const std::string& parametr, const std::vector<Polygon>& polys);
 size_t countEvenOdd(const std::string& parametr, const std::vector<Polygon>& polys);
-size_t countNumOfVertexes(int vertexes, const std::vector<Polygon>& polys);
+int countNumOfVertexes(int vertexes, const std::vector<Polygon>& polys);
 auto getBoundingBox(const std::vector<Polygon>& polys, int& minX,
   int& maxX, int& minY, int& maxY);
 bool inFrame(const std::vector<Polygon>& polys, const Polygon& target);
